% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.hat.R
\docType{data}
\name{cov.hat}
\alias{cov.hat}
\title{A sample covariance matrix from the same distribution as t.hat and Xs}
\format{A 200x200 data frame}
\usage{
cov.hat
}
\description{
The population precrision matrix has 20 connected components in the induced graph
}
\keyword{datasets}

