% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{GraphKit}
\alias{GraphKit}
\alias{GraphKit-package}
\title{GraphKit}
\description{
Efficient methods for constructing confidence intervals of monotone 
graph invariants, as well as testing for monotone graph properties. 
Many packages are available to estimate precision matrices, this 
package serves as a tool to extract structural properties from 
their induced graphs. By iteratively bootstrapping on only the 
relevant edge set, we are able to obtain the optimal interval size.
}
\author{
Manbir Gulati <manbirgulati@gmail.com>, Junwei Lu, Han Liu
}

