% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.hat.R
\docType{data}
\name{t.hat}
\alias{t.hat}
\title{A sample precision matrix from the same distribution as Xs and cov.hat}
\format{a 200x200 data frame}
\usage{
t.hat
}
\description{
The population precrision matrix has 20 connected components in the induced graph
}
\keyword{datasets}
