% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xs.R
\docType{data}
\name{Xs}
\alias{Xs}
\title{A random sample generated from a multivariate Guassian Distribution}
\format{A Large 500x200 Matrix}
\usage{
Xs
}
\description{
The population precrision matrix has 20 connected components in the induced graph
}
\keyword{datasets}
